Update ASP.NET Api to make it use Sql Server for the Database instead of Sqlite, and updated the docker files to reflect the new changes
