+++ server/Todo.Api/appsettings.Development.json
@@ -1,20 +1,22 @@
 {
-  "ConnectionStrings": {
-    "SqliteConnection": "Data Source=Database/todo.db",
-    "RedisConnection": "localhost:6379"
-  },
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Information",
-      "Microsoft.AspNetCore.HttpLogging.HttpLoggingMiddleware": "Information"
-    }
-  },
-  "JwtConfigurations": {
-    "SecretKey": "zd.7isDxc1#q{O2?8hpm<!CS9AUdP}>fJHfLmO+GCF),4s$Hi>zd.7isDxc1#q{O2?8hpm<!CS9AUdP}>fJHfLmO+GCF),4s$Hi>",
-    "Issuer": "Todo.Api",
-    "Audience": "http://localhost:5076",
-    "AccessTokenExpirationDays": 30,
-    "RefreshTokenExpirationDays": 90
-  }
+	"ConnectionStrings": {
+		"SqliteConnection": "Data Source=Database/todo.db",
+		"SqlServerConnection": "Server=sql.bsite.net\\MSSQL2016;Database=aladawy_todo_fullstack;User Id=aladawy_todo_fullstack;Password=Al-Adawy@123;trustServerCertificate=true;",
+		"RedisConnection": "localhost:6379"
+	},
+	"Logging": {
+		"LogLevel": {
+			"Default": "Information",
+			"Microsoft.AspNetCore": "Information",
+			"Microsoft.AspNetCore.HttpLogging.HttpLoggingMiddleware": "Information"
+		}
+	},
+	"JwtConfigurations": {
+		"SecretKey": "zd.7isDxc1#q{O2?8hpm<!CS9AUdP}>fJHfLmO+GCF),4s$Hi>zd.7isDxc1#q{O2?8hpm<!CS9AUdP}>fJHfLmO+GCF),4s$Hi>",
+		"Issuer": "Todo.Api",
+		"Audience": "http://localhost:5076",
+		"AccessTokenExpirationDays": 30,
+		"RefreshTokenExpirationDays": 90
+	},
+	"AllowedHosts": "*"
 }
