+++ server/Todo.Api/appsettings.json
@@ -1,6 +1,7 @@
 {
 	"ConnectionStrings": {
 		"SqliteConnection": "Data Source=Database/todo.db",
+		"SqlServerConnection": "Server=sql.bsite.net\\MSSQL2016;Database=aladawy_todo_fullstack;User Id=aladawy_todo_fullstack;Password=Al-Adawy@123;trustServerCertificate=true;",
 		"RedisConnection": "https://redis.todo.aladawy.duckdns.org"
 	},
 	"Logging": {
@@ -12,8 +13,8 @@
 	},
 	"JwtConfigurations": {
 		"SecretKey": "zd.7isDxc1#q{O2?8hpm<!CS9AUdP}>fJHfLmO+GCF),4s$Hi>zd.7isDxc1#q{O2?8hpm<!CS9AUdP}>fJHfLmO+GCF),4s$Hi>",
-		"Issuer": "Todo.Api",
-		"Audience": "http://localhost:5076",
+		"Issuer": "https://api.todo.aladawy.duckdns.org",
+		"Audience": "https://todo.aladawy.duckdns.org",
 		"AccessTokenExpirationDays": 30,
 		"RefreshTokenExpirationDays": 90
 	},
